+++ src/config/app/app.routes.ts
@@ -2,6 +2,7 @@
 const usersRoot = 'users';
 const authRoot = 'auth';
 const mailRoot = 'mail';
+const interestsRoot = 'interests';
 
 // Api Versions
 const v1 = 'v1';
@@ -25,6 +26,14 @@ export const routesV1 = {
     google: `/${authRoot}/google`,
     github: `/${authRoot}/github`,
     facebook: `/${authRoot}/facebook`,
+  },
+  interests: {
+    root: interestsRoot,
+    crate: `/${interestsRoot}`,
+    findAll: `/${interestsRoot}`,
+    delete: `/${interestsRoot}/:id`,
+    findOne: `/${interestsRoot}/:id`,
+    update: `/${interestsRoot}/:id`,
   },
   mail: {
     confirm: `/${mailRoot}/confirm`,