+++ src/infra/swagger/swagger.config.ts
@@ -14,6 +14,23 @@ API for QTalk
 ${changelog}
 `,
     )
-    .addBearerAuth()
+    .addBearerAuth(
+      {
+        type: 'http',
+        scheme: 'bearer',
+        bearerFormat: 'JWT',
+        description: 'Paste access token here',
+      },
+      'access-token',
+    )
+    .addBearerAuth(
+      {
+        type: 'http',
+        scheme: 'bearer',
+        bearerFormat: 'JWT',
+        description: 'Paste access token here',
+      },
+      'refresh-token',
+    )
     .build();
 }